Hot fix: In the listener for the network event, the `encoded` parameter in the callback is `dynamic`; expected: `int`
`encoded` casted as `int`

argument_type_no_assignable: The argument type 'dynamic' can't be assigned to the parameter type 'int'

+++ lib/ouisync_plugin.dart
@@ -140,7 +140,7 @@ class Session {
     final recvPort = ReceivePort();
 
     recvPort.listen((encoded) {
-      final event = _decodeNetworkEvent(encoded);
+      final event = _decodeNetworkEvent(encoded as int);
       if (event != null) {
         callback(event);
       } else {
